fix CG in func and main

+++ danilenko.ekaterina/I4/countColumnsWithAscElements.cpp
@@ -5,7 +5,7 @@ size_t countColumnsWithAscElements(const int* array, size_t rows, size_t columns
   size_t cols_count = columns;
   for (size_t j = 0; j < rows; j++)
   {
-    size_t prev_val = array[j * columns];
+    int prev_val = array[j * columns];
     for (size_t i = 1; i < columns; i++)
     {
       if (prev_val >= array[i + j * columns])
